cloud-api/requirements.txt
fastapi>=0.128.0
uvicorn>=0.39.0
python-multipart>=0.0.20
numpy>=2.0.0
opencv-python-headless>=4.10.0
insightface>=0.7.3
onnxruntime>=1.19.0
pydantic-settings>=2.7.0
